Continuation of the learning journal
In this update, I have started to add more content to the learning journal which is the Index page for the site.
In addition to this, I have started to add comments to the CSS in order to section it off so that it is easier to understand where everything should go, so in the next coming commits the CSS should be ordered correctly so that it is easier to find sections, and more additions to the index page will come.