CSS edit and HTML edit and post reorder
In this push, I have added borders to the posts Class so that users can tell where posts begin and end, with the help of my Classmate

The borders on the side have been removed so that the entire page is used rather than the middle third of the screen, but the website should also still scale when being used on a mobile device:
To also add better scaling to mobile devices I have made it so that the table and the navigation scale better, the table will disappear when the website is being used on mobile and the navigation bar will appear underneath the main Heading of the page.

The font Catamaran is now being used as the font for the page.
reference to the font: https://fonts.google.com/specimen/Catamaran?category=Sans+Serif,Display,Handwriting,Monospace

Posts on the actual website are now in descending order starting with the latest post and going down to the first post.

More HTML content:
references

@KennedySovine helped with CSS :)

Co-Authored-By: KennedySovine <89476714+KennedySovine@users.noreply.github.com>