Content addition, image adjustments, CSS changes and additions
Content for the next weeks are being added, changed some CSS so that it works correctly with the images and table, deleted MDN_Web_Docs.png because I realized it is not under creative commons licensing and that is naughty naughty.

In terms of the images that are still on the page, they were imported into GIMP and I changed the resolution on them so that they fit correctly in their respective places on the page.

CSS was added to create borders around the navigation options, and changes to how the links on the pages are displayed as well as changes to the text color of the navigation text and the header.

Added references from today into the references section of the page.

References:
https://www.w3schools.com/css/css3_borders.asp
https://www.december.com/html/spec/colorshades.html
https://www.w3schools.com/html/html_links_colors.asp