More content
Decided image banners not a good idea, so got rid of them, completed week 7 of the journal, only 2 more to complete along with some extra on the contact me page and the tutorial page needs to be done.
Did some more CSS work, not long left till this project is completed for the semester.